+++ sphinx/directives/other.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING, cast
 
 from docutils import nodes
 from docutils.parsers.rst import directives
-from docutils.parsers.rst.directives.admonitions import BaseAdmonition
 from docutils.parsers.rst.directives.misc import Class
 from docutils.parsers.rst.directives.misc import Include as BaseInclude
 from docutils.statemachine import StateMachine
@@ -213,12 +212,6 @@ class Author(SphinxDirective):
         ret: list[Node] = [para]
         ret += messages
         return ret
-
-
-class SeeAlso(BaseAdmonition):
-    """An admonition mentioning things to look at as reference."""
-
-    node_class = addnodes.seealso
 
 
 class TabularColumns(SphinxDirective):
@@ -427,7 +420,6 @@ def setup(app: Sphinx) -> ExtensionMetadata:
     directives.register_directive('sectionauthor', Author)
     directives.register_directive('moduleauthor', Author)
     directives.register_directive('codeauthor', Author)
-    directives.register_directive('seealso', SeeAlso)
     directives.register_directive('tabularcolumns', TabularColumns)
     directives.register_directive('centered', Centered)
     directives.register_directive('acks', Acks)