+++ sphinx/ext/todo.py
@@ -12,11 +12,10 @@ import operator
 from typing import TYPE_CHECKING, cast
 
 from docutils import nodes
-from docutils.parsers.rst import directives
-from docutils.parsers.rst.directives.admonitions import BaseAdmonition
 
 import sphinx
 from sphinx import addnodes
+from sphinx.directives.admonitions import SphinxAdmonition
 from sphinx.domains import Domain
 from sphinx.errors import NoUri
 from sphinx.locale import _, __
@@ -46,35 +45,25 @@ class todolist(nodes.General, nodes.Element):
     pass
 
 
-class Todo(BaseAdmonition, SphinxDirective):
+class Todo(SphinxAdmonition):
     """A todo entry, displayed (if configured) in the form of an admonition."""
 
     node_class = todo_node
-    has_content = True
-    required_arguments = 0
-    optional_arguments = 0
-    final_argument_whitespace = False
-    option_spec: ClassVar[OptionSpec] = {
-        'class': directives.class_option,
-        'name': directives.unchanged,
-    }
 
     def run(self) -> list[Node]:
         if not self.options.get('class'):
             self.options['class'] = ['admonition-todo']
 
         (todo,) = super().run()
-        if isinstance(todo, nodes.system_message):
+        if not isinstance(todo, todo_node):
             return [todo]
-        elif isinstance(todo, todo_node):
-            todo.insert(0, nodes.title(text=_('Todo')))
-            todo['docname'] = self.env.docname
-            self.add_name(todo)
-            self.set_source_info(todo)
-            self.state.document.note_explicit_target(todo)
-            return [todo]
-        else:
-            raise TypeError  # never reached here
+
+        todo.insert(0, nodes.title(text=_('Todo')))
+        todo['docname'] = self.env.docname
+        self.add_name(todo)
+        self.set_source_info(todo)
+        self.state.document.note_explicit_target(todo)
+        return [todo]
 
 
 class TodoDomain(Domain):